Refactored Engine Code
The engine has been completely refactored and engine masking was added, as alluded in the last commit. The engine now drives gameplay, instead of outside files. This allows for cleaner state updates, bug catches, etc. Currently in play-testing for bug catching since random inputs are not working (can't say why). Also removed the "ai_player.py" file as it no longer has any use in this version of the poker engine. "train_ai.py" has not been altered so its code is not compatible with the current "engine.py" code.